assigning files to correct folders
